Fix intptr_t/int32_t casting problem on Win64.
R=jkummerow@chromium.org

Review URL: https://codereview.chromium.org/11369133

git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@12899 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/x64/assembler-x64-inl.h
@@ -223,7 +223,7 @@ void RelocInfo::apply(intptr_t delta) {
   } else if (rmode_ == CODE_AGE_SEQUENCE) {
     if (*pc_ == kCallOpcode) {
       int32_t* p = reinterpret_cast<int32_t*>(pc_ + 1);
-      *p -= delta;  // Relocate entry.
+      *p -= static_cast<int32_t>(delta);  // Relocate entry.
       CPU::FlushICache(p, sizeof(uint32_t));
     }
   }
